requirements.txt
pytest==7.4.3
selenium==4.14.0
pytest-html==4.1.1
pandas==2.1.3
Unidecode==1.3.7
pymysql==1.1.0
mysql-connector-python==8.2.0
openpyxl==3.1.2
python-dotenv==1.0.0

patool==4.0.1
numpy==1.26.1